refactor: flatten ide scroll view structure
