+++ src/pages/shared/features/CodeViewer/model/atoms.ts
@@ -3,12 +3,12 @@
  * Tracks hovered file for right panel highlighting
  */
 
-import { atom } from 'jotai';
+import { atomWithDefault } from 'jotai/utils';
 
 /**
  * Currently hovered file path (via mouse hover)
  * Used to highlight the corresponding file in the right panel
  */
-export const hoveredFilePathAtom = atom<string | null>(null);
+export const hoveredFilePathAtom = atomWithDefault<string | null>(() => null);
 // IDE 모드에서 현재 포커스된 노드 ID
-export const focusedNodeIdAtom = atom<string | null>(null);
+export const focusedNodeIdAtom = atomWithDefault<string | null>(() => null);
