+++ src/pages/shared/features/OpenFiles/lib/useScrollNavigation.ts
