+++ src/pages/shared/features/OpenFiles/ui/FileNavPanel.tsx
