+++ src/pages/shared/features/OpenFiles/ui/FileSection.tsx
